Rename IDAFunctionParameterEstimation to ParameterEstimation
