+++ qt/scientific_interfaces/Inelastic/Analysis/FunctionBrowser/MSDFunctionModel.cpp
@@ -7,7 +7,7 @@
 #include "MSDFunctionModel.h"
 
 #include "Analysis/FitTabConstants.h"
-#include "Analysis/IDAFunctionParameterEstimation.h"
+#include "Analysis/ParameterEstimation.h"
 
 #include <memory>
