+++ qt/scientific_interfaces/Inelastic/Analysis/FunctionBrowser/SingleFunctionTemplateModel.cpp
@@ -5,7 +5,7 @@
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
 #include "SingleFunctionTemplateModel.h"
-#include "Analysis/IDAFunctionParameterEstimation.h"
+#include "Analysis/ParameterEstimation.h"
 #include "MantidAPI/FunctionFactory.h"
 #include "MantidAPI/IFunction.h"
 #include "MantidAPI/ITableWorkspace.h"
