+++ qt/scientific_interfaces/Inelastic/Analysis/FunctionBrowser/SingleFunctionTemplateView.cpp
@@ -5,7 +5,7 @@
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
 #include "SingleFunctionTemplateView.h"
-#include "Analysis/IDAFunctionParameterEstimation.h"
+#include "Analysis/ParameterEstimation.h"
 
 #include "MantidAPI/CostFunctionFactory.h"
 #include "MantidAPI/FuncMinimizerFactory.h"
