+++ qt/scientific_interfaces/Inelastic/Analysis/ParameterEstimation.cpp
@@ -4,7 +4,7 @@
 //   NScD Oak Ridge National Laboratory, European Spallation Source,
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
-#include "IDAFunctionParameterEstimation.h"
+#include "ParameterEstimation.h"
 #include "MantidAPI/IFunction.h"
 
 #include <math.h>
